うまく分割できなかったので元に戻した
