--- console.go
@@ -1,15 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"log"
-)
-
-func main2() {
-	log.Println("Start")
-
-	message := buildMessage()
-	fmt.Println(message)
-
-	log.Println("Finish")
-}
